organized file structure in prep for more complexity
