+++ kale-yeah-web/src/App.js
@@ -2,13 +2,16 @@ import { Route, Routes } from 'react-router-dom';
 
 import './index.css';
 import LandingPage from './pages/Landing';
+import Login from './components/Login';
 
 function App() {
   return (
     <div>
-      <Routes>
+      {/* <Routes>
         <Route path='/' element={<LandingPage />} exact/>
-      </Routes>
+      </Routes> */}
+  
+      <Login />
     </div>
   );
