+++ kale-yeah-web/src/index.css
@@ -24,3 +24,171 @@
   padding: 5px;
   margin: 5px;
 }
+
+
+
+.ky {
+  position: relative;
+  top: -40px;
+  font-family: 'Courier New', Courier, monospace;
+  color: #3eb489;
+}
+
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  padding: 2rem;
+}
+
+.btn {
+  background-color: #3eb489;
+  border: 1px solid #3eb489;
+  border-radius: 4px;
+  box-shadow: rgba(0, 0, 0, .1) 0 2px 4px 0;
+  box-sizing: border-box;
+  color: #fff;
+  cursor: pointer;
+  font-family: "Akzidenz Grotesk BQ Medium", -apple-system, BlinkMacSystemFont, sans-serif;
+  font-size: 16px;
+  font-weight: 400;
+  outline: none;
+  outline: 0;
+  padding: 10px 25px;
+  text-align: center;
+  transform: translateY(0);
+  transition: transform 150ms, box-shadow 150ms;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+}
+
+.btn:hover {
+  box-shadow: rgba(0, 0, 0, .15) 0 3px 9px 0;
+  transform: translateY(-2px);
+}
+
+@media (min-width: 768px) {
+  button {
+    padding: 10px 30px;
+  }
+}
+
+
+
+.form-container {
+  background: #fff;
+  width: 300px;
+  height: 450px;
+  position: relative;
+  top: -18px;
+  text-align: center;
+  padding: 20px 0;
+  margin: auto;
+  box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+}
+
+.form-btn {
+  display: inline-block;
+}
+
+.form-container form {
+  max-width: 300px;
+  padding: 0 20px;
+  position: absolute;
+  top: 120px;
+  transition: transform 1s;
+}
+
+form input {
+  width: 100%;
+  height: 30px;
+  margin: 10px 0;
+  padding: 0 10px;
+  border: 1px solid #ccc;
+}
+
+form .btn {
+  color: white;
+  width: 100%;
+  border: none;
+  cursor: pointer;
+  margin: 10px 0;
+}
+
+form .btn:focus {
+  outline: none;
+}
+
+#loginform {
+  left: 0px;
+}
+
+#Regform {
+  left: 0px;
+}
+
+form a {
+  font-size: 14px;
+}
+
+input[type=email]:not(:placeholder-shown):invalid {
+  color: black;
+}
+
+ #login{
+   background: none;
+   background-image: none;
+   border: none;
+   margin:0;
+   padding:0;
+   cursor: pointer;
+   font-size: medium;
+   font-weight: bold;
+   padding: 0 10px;
+   color: #555;
+   width: 100px;
+   display: inline-block;
+   background-color: Transparent;
+   background-repeat:no-repeat;
+   overflow: hidden;
+
+ } 
+
+#login:hover{
+  color: #3eb489;
+}
+
+#login:focus{
+  color: #3eb489;
+}
+
+#register:hover {
+  color: #3eb489;
+
+}
+
+#register:focus{
+  color: #3eb489;
+}
+  
+  
+ #register{
+  background: none;
+  border: none;
+  margin:0;
+  padding:0;
+  font-size: medium;
+  font-weight: bold;
+  padding: 0 10px;
+  color: #555;
+  cursor: pointer;
+  width: 100px;
+  display: inline-block;
+
+ }
